+++ test/unit/fuzz-tools.test.js
@@ -273,6 +273,7 @@ import * as transfusionV292 from '../../lib/transfusion-v292.js';
 import * as fastDementiaV294 from '../../lib/fast-dementia-v294.js';
 import * as gdsV295 from '../../lib/gds-v295.js';
 import * as benzoEquivV296 from '../../lib/benzo-equiv-v296.js';
+import * as nerveInjuryV297 from '../../lib/nerve-injury-v297.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -491,6 +492,7 @@ const MODULES = {
   'fast-dementia-v294.js': fastDementiaV294,
   'gds-v295.js': gdsV295,
   'benzo-equiv-v296.js': benzoEquivV296,
+  'nerve-injury-v297.js': nerveInjuryV297,
   'tb-testing.js': tbTesting,
 };